Moved all mocks to separate files
