+++ TimeTable/Models/BussinessModels/WorkTimes/ApiClient+WorkTimes.swift
@@ -1,0 +1,19 @@
+//
+//  ApiClient+WorkTimes.swift
+//  TimeTable
+//
+//  Created by Piotr Pawluś on 22/11/2018.
+//  Copyright © 2018 Railwaymen. All rights reserved.
+//
+
+import Foundation
+
+protocol ApiClientWorkTimesType: class {
+    func fetchWorkTimes(parameters: WorkTimesPrameters, completion: @escaping ((Result<[WorkTimeDecoder]>) -> Void))
+}
+
+extension ApiClient: ApiClientWorkTimesType {
+    func fetchWorkTimes(parameters: WorkTimesPrameters, completion: @escaping ((Result<[WorkTimeDecoder]>) -> Void)) {
+        get(.worktimes, parameters: parameters, completion: completion)
+    }
+}
