+++ TimeTableTests/Categories/Login/LoginContentProviderTests.swift
@@ -159,64 +159,3 @@ private class ApiClientSessionMock: ApiClientSessionType {
         signInCompletion = completion
     }
 }
-
-private class CoreDataStackUserMock: CoreDataStackUserType {
-    
-    typealias CDT = UserEntity
-    private(set) var fetchUserIdentifier: Int?
-    private(set) var fetchUserCompletion: ((Result<UserEntity>) -> Void)?
-    private(set) var saveUserDecoder: SessionDecoder?
-    private(set) var saveUserCompletion: ((Result<UserEntity>) -> Void)?
-    private(set) var saveCoreDataTypeTranslatior: ((AsynchronousDataTransactionType) -> UserEntity)?
-    
-    func fetchUser(forIdentifier identifier: Int, completion: @escaping (Result<UserEntity>) -> Void) {
-        fetchUserIdentifier = identifier
-        fetchUserCompletion = completion
-    }
-    
-    func save<CDT>(userDecoder: SessionDecoder,
-                   coreDataTypeTranslation: @escaping ((AsynchronousDataTransactionType) -> CDT),
-                   completion: @escaping (Result<CDT>) -> Void) {
-    
-        // swiftlint:disable force_cast
-        saveUserDecoder = userDecoder
-        saveUserCompletion = { result in
-            switch result {
-            case .failure(let error):
-                completion(.failure(error))
-            case .success(let entity):
-                completion(.success(entity as! CDT))
-            }
-        }
-        saveCoreDataTypeTranslatior = { transaction in
-            return coreDataTypeTranslation(transaction) as! UserEntity
-        }
-        // swiftlint:enable force_cast
-    }
-}
-
-private class AsynchronousDataTransactionMock: AsynchronousDataTransactionType {
-    
-    private(set) var deleteAllCalled = false
-    private(set) var createCalled = false
-    var user: DynamicObject?
-    
-    func deleteAll<D>(_ from: From<D>, _ deleteClauses: DeleteClause...) -> Int? where D: DynamicObject {
-        deleteAllCalled = true
-        return nil
-    }
-    
-    func create<D>(_ into: Into<D>) -> D where D: DynamicObject {
-        createCalled = true
-        // swiftlint:disable force_cast
-        return user as! D
-        // swiftlint:enable force_cast
-    }
-}
-
-private class AccessServiceMock: AccessServiceUserIDType {
-    func saveLastLoggedInUserIdentifier(_ identifer: Int64) {}
-    func getLastLoggedInUserIdentifier() -> Int64? {
-        return nil
-    }
-}
