+++ TimeTableTests/Coordinators/AppCoordinatorTests.swift
@@ -218,67 +218,3 @@ class AppCoordinatorTests: XCTestCase {
         )
     }
 }
-
-private class StoryboardsManagerMock: StoryboardsManagerType {
-    var controller: UIViewController?
-    func controller<T>(storyboard: StoryboardsManager.StoryboardName, controllerIdentifier: StoryboardsManager.ControllerIdentifier) -> T? {
-        return controller as? T
-    }
-}
-
-private class ErrorHandlerMock: ErrorHandlerType {
-    func catchingError(action: @escaping (Error) throws -> Void) -> ErrorHandlerType {
-        return ErrorHandler(action: action)
-    }
-    
-    func throwing(error: Error, finally: @escaping (Bool) -> Void) {}
-}
-
-private class ServerConfigurationViewControllerMock: ServerConfigurationViewControlleralbe {
-    func configure(viewModel: ServerConfigurationViewModelType, notificationCenter: NotificationCenterType) {}
-    func setupView(checkBoxIsActive: Bool, serverAddress: String) {}
-    func tearDown() {}
-    func hideNavigationBar() {}
-    func continueButtonEnabledState(_ isEnabled: Bool) {}
-    func checkBoxIsActiveState(_ isActive: Bool) {}
-    func dissmissKeyboard() {}
-}
-
-private class LoginViewControllerMock: LoginViewControllerable {
-    func configure(notificationCenter: NotificationCenterType, viewModel: LoginViewModelType) {}
-    func setUpView(checkBoxIsActive: Bool) {}
-    func updateLoginFields(email: String, password: String) {}
-    func tearDown() {}
-    func passwordInputEnabledState(_ isEnabled: Bool) {}
-    func loginButtonEnabledState(_ isEnabled: Bool) {}
-    func checkBoxIsActiveState(_ isActive: Bool) {}
-    func focusOnPasswordTextField() {}
-}
-
-private class ServerConfigurationManagerMock: ServerConfigurationManagerType {
-    var oldConfiguration: ServerConfiguration?
-    private(set) var oldConfigurationCalled = false
-    private(set) var verifyConfigurationValues: (called: Bool, configuration: ServerConfiguration?) = (false, nil)
-    private(set) var verifyConfigurationCompletion: ((Result<Void>) -> Void)?
-    
-    func getOldConfiguration() -> ServerConfiguration? {
-        oldConfigurationCalled = true
-        return oldConfiguration
-    }
-    
-    func verify(configuration: ServerConfiguration, completion: @escaping ((Result<Void>) -> Void)) {
-        verifyConfigurationValues = (true, configuration)
-        verifyConfigurationCompletion = completion
-    }
-}
-
-private class CoreDataStackMock: CoreDataStackType {
-    func save<CDT>(userDecoder: SessionDecoder,
-                   coreDataTypeTranslation: @escaping ((AsynchronousDataTransactionType) -> CDT),
-                   completion: @escaping (Result<CDT>) -> Void) where CDT: NSManagedObject {}
-    func fetchUser(forIdentifier identifier: Int, completion: @escaping (Result<UserEntity>) -> Void) {}
-}
-
-private class BundleMock: BundleType {
-    var bundleIdentifier: String?
-}
