+++ TimeTableTests/Coordinators/AuthenticationCoordinatorTests.swift
@@ -77,59 +77,3 @@ class AuthenticationCoordinatorTests: XCTestCase {
         XCTAssertTrue(finishCompletionCalled)
     }
 }
-
-private class StoryboardsManagerMock: StoryboardsManagerType {
-    var controller: UIViewController?
-    func controller<T>(storyboard: StoryboardsManager.StoryboardName, controllerIdentifier: StoryboardsManager.ControllerIdentifier) -> T? {
-        return controller as? T
-    }
-}
-
-private class ErrorHandlerMock: ErrorHandlerType {
-    func catchingError(action: @escaping (Error) throws -> Void) -> ErrorHandlerType {
-        return ErrorHandler(action: action)
-    }
-    
-    func throwing(error: Error, finally: @escaping (Bool) -> Void) {}
-}
-
-private class LoginViewControllerMock: LoginViewControllerable {
-    func configure(notificationCenter: NotificationCenterType, viewModel: LoginViewModelType) {}
-    func setUpView(checkBoxIsActive: Bool) {}
-    func updateLoginFields(email: String, password: String) {}    
-    func tearDown() {}
-    func passwordInputEnabledState(_ isEnabled: Bool) {}
-    func loginButtonEnabledState(_ isEnabled: Bool) {}
-    func checkBoxIsActiveState(_ isActive: Bool) {}
-    func focusOnPasswordTextField() {}
-}
-
-private class ApiClientMock: ApiClientSessionType {
-    
-    private(set) var signInCredentials: LoginCredentials?
-    private(set) var signInCompletion: ((Result<SessionDecoder>) -> Void)?
-    
-    func signIn(with credentials: LoginCredentials, completion: @escaping ((Result<SessionDecoder>) -> Void)) {
-        signInCredentials = credentials
-        signInCompletion = completion
-    }
-}
-
-private class CoreDataStackMock: CoreDataStackType {
-    func save<CDT>(userDecoder: SessionDecoder,
-                   coreDataTypeTranslation: @escaping ((AsynchronousDataTransactionType) -> CDT),
-                   completion: @escaping (Result<CDT>) -> Void) where CDT: NSManagedObject {}
-    func fetchUser(forIdentifier identifier: Int, completion: @escaping (Result<UserEntity>) -> Void) {}
-}
-
-private class AccessServiceMock: AccessServiceLoginType {
-    func saveLastLoggedInUserIdentifier(_ identifer: Int64) {}
-    func getLastLoggedInUserIdentifier() -> Int64? {
-        return nil
-    }
-    func saveUser(credentails: LoginCredentials) throws {}
-    func getUserCredentials() throws -> LoginCredentials {
-        return LoginCredentials(email: "", password: "")
-    }
-    func removeLastLoggedInUserIdentifier() {}
-}
