+++ TimeTableTests/Coordinators/Base/BaseCoordinatorTests.swift
@@ -185,13 +185,3 @@ private class ChildCoordinator: BaseCoordinator {
         super.finish()
     }
 }
-
-private struct TestError: Error {
-    let messsage: String
-}
-
-extension TestError: Equatable {
-    static func == (lhs: TestError, rhs: TestError) -> Bool {
-        return lhs.messsage == rhs.messsage
-    }
-}
