+++ TimeTableTests/Mocks/BundleMock.swift
@@ -1,0 +1,14 @@
+//
+//  BundleMock.swift
+//  TimeTableTests
+//
+//  Created by Piotr Pawluś on 22/11/2018.
+//  Copyright © 2018 Railwaymen. All rights reserved.
+//
+
+import Foundation
+@testable import TimeTable
+
+class BundleMock: BundleType {
+    var bundleIdentifier: String?
+}
