+++ TimeTableTests/Mocks/StoryboardsManagerMock.swift
@@ -1,0 +1,17 @@
+//
+//  StoryboardsManagerMock.swift
+//  TimeTableTests
+//
+//  Created by Piotr Pawluś on 22/11/2018.
+//  Copyright © 2018 Railwaymen. All rights reserved.
+//
+
+import UIKit
+@testable import TimeTable
+
+class StoryboardsManagerMock: StoryboardsManagerType {
+    var controller: UIViewController?
+    func controller<T>(storyboard: StoryboardsManager.StoryboardName, controllerIdentifier: StoryboardsManager.ControllerIdentifier) -> T? {
+        return controller as? T
+    }
+}
