+++ TimeTableTests/Models/AccessService/AccessServiceTests.swift
@@ -112,7 +112,7 @@ class AccessServiceTests: XCTestCase {
             _ = try accessService.getUserCredentials()
         } catch {
             //Assert
-            XCTAssertEqual(try (error as? TestError).unwrap(), TestError(messsage: "decode error"))
+            XCTAssertEqual(try (error as? TestError).unwrap(), TestError(messsage: "decoder error"))
         }
     }
 
@@ -137,7 +137,7 @@ class AccessServiceTests: XCTestCase {
         //Act
         accessService.saveLastLoggedInUserIdentifier(identifier)
         //Assert
-        XCTAssertEqual(try (userDefaultsMock.setAnyValue as? Int64).unwrap(), identifier)
+        XCTAssertEqual(try (userDefaultsMock.setAnyValues.value as? Int64).unwrap(), identifier)
     }
     
     func testGetLastLoggedInUserIdentifierReturnsNilWhileUserDefaultsReturnsNil() {
@@ -172,98 +172,6 @@ class AccessServiceTests: XCTestCase {
         //Act
         accessService.removeLastLoggedInUserIdentifier()
         //Assert
-        XCTAssertTrue(userDefaultsMock.removeObjectCalled)
-    }
-}
-
-private class UserDefaultsMock: UserDefaultsType {
-    private(set) var removeObjectCalled = false
-    private(set) var setAnyValue: Any?
-    var objectForKey: Any?
-    
-    func bool(forKey defaultName: String) -> Bool { return false }
-    
-    func removeObject(forKey defaultName: String) {
-        removeObjectCalled = true
-    }
-    
-    func set(_ value: Any?, forKey defaultName: String) {
-        self.setAnyValue = value
-    }
-    
-    func set(_ value: Bool, forKey defaultName: String) {}
-    func string(forKey defaultName: String) -> String? { return nil }
-    
-    func object(forKey defaultName: String) -> Any? {
-        return objectForKey
-    }
-}
-
-private class KeychainAccessMock: KeychainAccessType {
-    var setDataIsThrowingError = false
-    var getDataIsThrowingError = false
-    var getDataValue: Data?
-    
-    func get(_ key: String) throws -> String? {
-        return nil
-    }
-    func getData(_ key: String) throws -> Data? {
-        if getDataIsThrowingError {
-            throw TestError(messsage: "set Data error")
-        } else {
-            return getDataValue
-        }
-    }
-    func set(_ value: String, key: String) throws {}
-
-    func set(_ value: Data, key: String) throws {
-        if setDataIsThrowingError {
-            throw TestError(messsage: "set Data error")
-        }
-    }
-}
-
-private class JSONEncoderMock: JSONEncoderType {
-    private lazy var encoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = .prettyPrinted
-        return encoder
-    }()
-    
-    var isThrowingError = false
-    
-    var dateEncodingStrategy: JSONEncoder.DateEncodingStrategy = .iso8601
-    func encode<T>(_ value: T) throws -> Data where T: Encodable {
-        if isThrowingError {
-            throw TestError(messsage: "encode error")
-        } else {
-            return try encoder.encode(value)
-        }
-    }
-}
-
-private class JSONDecoderMock: JSONDecoderType {
-    
-    private lazy var decoder: JSONDecoder = JSONDecoder()
-    
-    var isThrowingError = false
-    var dateDecodingStrategy: JSONDecoder.DateDecodingStrategy = .iso8601
-    
-    func decode<T>(_ type: T.Type, from data: Data) throws -> T where T: Decodable {
-        if isThrowingError {
-            throw TestError(messsage: "decode error")
-        } else {
-            return try decoder.decode(T.self, from: data)
-        }
-    }
-}
-
-private struct TestError: Error {
-    let messsage: String
-}
-
-extension TestError: Equatable {
-    static func == (lhs: TestError, rhs: TestError) -> Bool {
-        return lhs.messsage == rhs.messsage
+        XCTAssertTrue(userDefaultsMock.removeObjectValues.called)
     }
 }
