+++ TimeTableTests/Models/ApiClient/ApiClientTests.swift
@@ -133,135 +133,4 @@ class ApiClientTests: XCTestCase {
         default: XCTFail()
         }
     }
-    
-    // MARK: - ApiClientSessionType
-    func testSignInSucced() throws {
-        //Arrange
-        let data = try self.json(from: SessionResponse.signInResponse)
-        let decoder = try JSONDecoder().decode(SessionDecoder.self, from: data)
-        var expecdedSessionDecoder: SessionDecoder?
-        let parameters = LoginCredentials(email: "user1@example.com", password: "password")
-        let apiClient: ApiClientSessionType = ApiClient(networking: networkingMock, buildEncoder: { () -> RequestEncoderType in
-            return requestEncoderMock
-        }) { () -> JSONDecoderType in
-            return jsonDecoderMock
-        }
-        //Act
-        apiClient.signIn(with: parameters) { result in
-            switch result {
-            case .success(let sessionDecoder):
-                expecdedSessionDecoder = sessionDecoder
-            case .failure:
-                XCTFail()
-            }
-        }
-        networkingMock.shortPostCompletion?(.success(data))
-        //Assert
-        XCTAssertEqual(try expecdedSessionDecoder.unwrap(), decoder)
-    }
-    
-    func testSignInFailed() throws {
-        //Arrange
-        var expecdedError: Error?
-        let error = TestError(messsage: "sign in failed")
-        let parameters = LoginCredentials(email: "user1@example.com", password: "password")
-        let apiClient: ApiClientSessionType = ApiClient(networking: networkingMock, buildEncoder: { () -> RequestEncoderType in
-            return requestEncoderMock
-        }) { () -> JSONDecoderType in
-            return jsonDecoderMock
-        }
-        //Act
-        apiClient.signIn(with: parameters) { result in
-            switch result {
-            case .success:
-                XCTFail()
-            case .failure(let error):
-                expecdedError = error
-            }
-        }
-        networkingMock.shortPostCompletion?(.failure(error))
-        //Assert
-        let testError = try (expecdedError as? TestError).unwrap()
-        XCTAssertEqual(testError, error)
-    }
-}
-
-private class NetworkingMock: NetworkingType {
-    private(set) var shortPostValues: (path: String, parametres: Any?)?
-    private(set) var shortPostCompletion: ((TimeTable.Result<Data>) -> Void)?
-    
-    var headerFields: [String: String]?
-    
-    func post(_ path: String, parameters: Any?, completion: @escaping (_ result: TimeTable.Result<Data>) -> Void) -> String {
-        shortPostValues = (path, parameters)
-        shortPostCompletion = completion
-        return ""
-    }
-}
-
-private class RequestEncoderMock: RequestEncoderType {
-    private lazy var encoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = .prettyPrinted
-        return encoder
-    }()
-    
-    private(set) var encodeWrapper: Encodable?
-    var isEncodeThrowingError = false
-
-    func encode<T>(wrapper: T) throws -> Data where T: Encodable {
-        self.encodeWrapper = wrapper
-        if isEncodeThrowingError {
-            throw TestError(messsage: "encode error")
-        } else {
-            return try encoder.encode(wrapper)
-        }
-    }
-    
-    private(set) var encodeToDictionaryWrapper: Encodable?
-    var isEncodeToDictionaryThrowingError = false
-    
-    func encodeToDictionary<T>(wrapper: T) throws -> [String: Any] where T: Encodable {
-        self.encodeToDictionaryWrapper = wrapper
-        if isEncodeToDictionaryThrowingError {
-            throw TestError(messsage: "encode to dictionary error")
-        } else {
-            let data = try encoder.encode(wrapper)
-            guard let json = try JSONSerialization.jsonObject(with: data, options: .allowFragments) as? [String: Any] else {
-                throw TestError(messsage: "JSONSerialization.jsonObject error")
-            }
-            return json
-        }
-    }
-}
-
-private class JSONDecoderMock: JSONDecoderType {
-    
-    private lazy var decoder = JSONDecoder()
-    
-    var isDecodeThrowingError = false
-    private(set) var decodeType: Decodable.Type?
-    private(set) var decodeData: Data?
-    
-    var dateDecodingStrategy: JSONDecoder.DateDecodingStrategy = .iso8601
-    
-    func decode<T>(_ type: T.Type, from data: Data) throws -> T where T: Decodable {
-        decodeType = type
-        decodeData = data
-        if isDecodeThrowingError {
-            throw TestError(messsage: "decoder error")
-        } else {
-            return try decoder.decode(type, from: data)
-        }
-    }
-}
-
-private struct TestError: Error {
-    let messsage: String
-}
-
-extension TestError: Equatable {
-    static func == (lhs: TestError, rhs: TestError) -> Bool {
-        return lhs.messsage == rhs.messsage
-    }
 }
