+++ TimeTableTests/Models/ApiClient/RequestEncoder.swift
@@ -94,47 +94,3 @@ class RequestEncoderTests: XCTestCase {
         }
     }
 }
-
-private class JSONEncoderMock: JSONEncoderType {
-    private lazy var encoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = .prettyPrinted
-        return encoder
-    }()
-    
-    var isThrowingError = false
-    
-    var dateEncodingStrategy: JSONEncoder.DateEncodingStrategy = .iso8601
-    func encode<T>(_ value: T) throws -> Data where T: Encodable {
-        if isThrowingError {
-            throw TestError(messsage: "encode error")
-        } else {
-            return try encoder.encode(value)
-        }
-    }
-}
-
-private class JSONSerializationMock: JSONSerializationType {
-    var isThrowingError = false
-    var customObject: Any?
-    
-    func jsonObject(with data: Data, options opt: JSONSerialization.ReadingOptions) throws -> Any {
-        if isThrowingError {
-            throw TestError(messsage: "jsonObject error")
-        } else if let object = customObject {
-            return object
-        } else {
-            return try JSONSerialization.jsonObject(with: data, options: opt)
-        }
-    }
-}
-
-private struct TestError: Error {
-    let messsage: String
-}
-
-extension TestError: Equatable {
-    static func == (lhs: TestError, rhs: TestError) -> Bool {
-        return lhs.messsage == rhs.messsage
-    }
-}
