+++ TimeTableTests/Models/BussinessModels/WorkTimes/ApiClientWorkTimesTests.swift
@@ -1,0 +1,14 @@
+//
+//  ApiClient+WorkTimes.swift
+//  TimeTableTests
+//
+//  Created by Piotr Pawluś on 22/11/2018.
+//  Copyright © 2018 Railwaymen. All rights reserved.
+//
+
+import XCTest
+@testable import TimeTable
+
+class ApiClientWorkTimesTests: XCTestCase {
+    
+}
