+++ TimeTableTests/Models/CoreData/Translators/UserEntitySessionDecoderTests.swift
@@ -7,7 +7,7 @@
 //
 
 import XCTest
-import CoreStore
+import CoreData
 @testable import TimeTable
 
 class UserEntitySessionDecoderTests: XCTestCase {
@@ -48,24 +48,5 @@ class UserEntitySessionDecoderTests: XCTestCase {
         XCTAssertEqual(createdUser.firstName, sessionReponse.firstName)
         XCTAssertEqual(createdUser.lastName, sessionReponse.lastName)
         XCTAssertEqual(createdUser.token, sessionReponse.token)
-    }
-}
-
-private class AsynchronousDataTransactionMock: AsynchronousDataTransactionType {
-
-    private(set) var deleteAllCalled = false
-    private(set) var createCalled = false
-    var user: DynamicObject?
-    
-    func deleteAll<D>(_ from: From<D>, _ deleteClauses: DeleteClause...) -> Int? where D: DynamicObject {
-        deleteAllCalled = true
-        return nil
-    }
-    
-    func create<D>(_ into: Into<D>) -> D where D: DynamicObject {
-        createCalled = true
-        // swiftlint:disable force_cast
-        return user as! D
-        // swiftlint:enable force_cast
     }
 }
