+++ TimeTableTests/Models/Errors/ErrorHanlderTests.swift
@@ -62,13 +62,3 @@ class ErrorHanlderTests: XCTestCase {
         
     }
 }
-
-private struct TestError: Error {
-    let messsage: String
-}
-
-extension TestError: Equatable {
-    static func == (lhs: TestError, rhs: TestError) -> Bool {
-        return lhs.messsage == rhs.messsage
-    }
-}
